Refactor pasapi users and CMD
