+++ pkg/cybr/api/requests/unsuspenduser.go
@@ -1,0 +1,6 @@
+package requests
+
+// UnsuspendUser request used when unsuspending user
+type UnsuspendUser struct {
+	Suspended bool `json:"Suspended"`
+}
